feat: add utils, errors to exports | reorganize structure
