+++ packages/sdk/src/errors.ts
