--- packages/sdk/src/provider/index.ts
@@ -1,2 +1,0 @@
-export * from "./object.js";
-export * from "./utils.js";
