+++ packages/sdk/tsup.config.ts
@@ -7,9 +7,11 @@ export default defineConfig({
     bucket: "src/entities/bucket.ts",
     client: "src/client.ts",
     credit: "src/entities/credit.ts",
+    errors: "src/errors.ts",
     ipc: "src/ipc/index.ts",
     network: "src/network.ts",
-    provider: "src/provider/index.ts",
+    provider: "src/provider.ts",
+    utils: "src/utils.ts",
   },
   format: ["cjs", "esm"],
   dts: true,
